Refactor GalleryImage into a component
