+++ src/GalleryImage.jsx
@@ -1,0 +1,41 @@
+import React, { PropTypes } from 'react';
+import styles   from './index.scss';
+import LazyLoad from 'react-lazyload';
+
+const GalleryImage = ({ image, index }) => {
+  return (
+    <div className={styles.section}>
+      <div className={styles.sidebar}>
+          <div
+              className={styles.headline}
+              data-aos="slide-up"
+              data-aos-offset="200"
+              >{image.title}</div>
+          <div
+              className={styles.subtitle}
+              data-aos="slide-up"
+              data-aos-offset="200"
+              >{image.subtitle}</div>
+          <div
+              className={styles.moreInfo}
+              data-aos="slide-up"
+              data-aos-offset="200"
+              >{image.info}</div>
+      </div>
+      <LazyLoad height={200} offset={300}>
+          <img
+              src={image.src}
+              data-aos="fade-in"
+              className={styles.image + ' ' + index}
+              ></img>
+      </LazyLoad>
+    </div>
+  )
+};
+
+GalleryImage.propTypes = {
+  image: PropTypes.object.isRequired,
+  index: PropTypes.number.isRequired,
+};
+
+export default GalleryImage;
