+++ src/app.jsx
@@ -3,7 +3,7 @@ import styles       from './index.scss';
 import React        from 'react';
 import Intro        from './Intro.jsx';
 import Gallery      from './Gallery.jsx';
-import GalleryImages from './GalleryImages.js';
+import selectedImages from './selectedImages.js';
 import ScrollArrow  from './ScrollArrow.jsx';
 import $            from 'jquery';
 
@@ -41,7 +41,7 @@ export default class App extends React.Component {
 			<div>
 				<Intro/>
                 { this.state.showScroll ? <ScrollArrow />: null }
-				<Gallery images={GalleryImages} />
+				<Gallery images={selectedImages} />
             </div>
         )
     }
